# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Access-Nex
  description: >
    Self-hosted OAuth2 / OpenID Connect provider. This document covers the
    JSON surface a client application (including a separately-hosted
    frontend — see docs/FRONTEND.md) can build against: the standard
    OIDC/OAuth2 endpoints, the /api/v1 account API, and the /api/admin
    management API. The server-rendered HTML pages (/login, /portal, /admin,
    /device) are not part of this contract — they exist for zero-setup
    browser use and can be replaced entirely by a custom frontend calling
    these APIs.
  version: "1.0"
servers:
  - url: http://localhost:8080
    description: Default local install
security:
  - sessionCookie: []
paths:
...
  /token:
    post:
      summary: Token endpoint
      security:
        - clientBasic: []
        - {}
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                grant_type:
                  type: string
                  enum:
                    - authorization_code
                    - refresh_token
                    - client_credentials
                    - urn:ietf:params:oauth:grant-type:device_code
                    - urn:ietf:params:oauth:grant-type:token-exchange
                code: { type: string }
                redirect_uri: { type: string }
                code_verifier: { type: string, description: "PKCE verifier" }
                refresh_token: { type: string }
                scope: { type: string }
                audience: { type: string, description: "Repeatable; extra audiences for the issued token" }
                device_code: { type: string }
                subject_token: { type: string, description: "Token exchange only" }
                subject_token_type: { type: string, description: "urn:ietf:params:oauth:token-type:access_token" }
                client_id: { type: string }
                client_secret: { type: string }
      parameters:
        - name: DPoP
          in: header
          required: false
          schema: { type: string }
          description: RFC 9449 sender-constrained token proof (optional)
      responses:
        "200":
          description: Token response
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token: { type: string }
                  token_type: { type: string, enum: [Bearer, DPoP] }
                  expires_in: { type: integer }
                  refresh_token: { type: string }
                  id_token: { type: string, description: "JWT, or a JWE compact string if the client registered an ID-token encryption key" }
                  scope: { type: string }
        "400":
          description: >
            OAuth error (invalid_grant, invalid_request, authorization_pending,
            slow_down, access_denied, expired_token, use_dpop_nonce, ...)
...
  /api/v1/login:
    post:
      summary: Password login (step 1)
      security: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties: { username: { type: string }, password: { type: string } }
      responses:
        "200":
          description: Signed in, or a 2FA challenge
          content:
            application/json:
              schema:
                oneOf:
                  - type: object
                    properties: { status: { type: string, enum: [signed_in] } }
                  - type: object
                    properties:
                      status: { type: string, enum: [2fa_required] }
                      token: { type: string, description: "pass to /api/v1/login/totp or /login/webauthn/*" }
                      methods: { type: array, items: { type: string, enum: [totp, webauthn] } }
        "401": { description: Invalid credentials }
        "429": { description: Rate limited }
...
  /api/v1/forgot-password:
    post:
      summary: Request a password reset email
      security: []
      requestBody:
        content:
          application/json:
            schema: { type: object, properties: { email: { type: string } } }
      responses:
        "200":
          description: >
            Always returns { status: if_exists_email_sent } regardless of
            whether the address matched an account, to avoid user enumeration.
  /api/v1/reset-password:
    post:
      summary: Complete a password reset (token from the emailed link)
      security: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties: { token: { type: string }, new_password: { type: string, minLength: 8 } }
      responses:
        "200": { description: Changed }
        "400": { description: Link invalid or expired }
...
components:
...
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: oidc_sid
      description: >
        Set by /login, /api/v1/login(+totp), or /login/webauthn/finish.
        Mutating /api/v1 and /api/admin requests additionally require the
        X-Access-Nex-Api / X-Access-Nex-Admin header respectively.
    clientBasic:
      type: http
      scheme: basic
      description: OAuth2 client_id:client_secret (confidential clients); public clients omit auth entirely
    bearerToken:
      type: http
      scheme: bearer
    dpopToken:
      type: apiKey
      in: header
      name: Authorization
      description: 'For DPoP-bound tokens: "Authorization: DPoP <token>" plus a DPoP proof header'